+++ src/sections/instructor/components/user-profile/_components/profile-section.tsx
@@ -1,4 +1,4 @@
-import { useRef, useState } from 'react'
+import React, { useRef, useState } from 'react'
 import { zodResolver } from '@hookform/resolvers/zod'
 import { Camera, Loader2, Mail, MapPin, Phone, User } from 'lucide-react'
 import { useForm } from 'react-hook-form'
