+++ src/sections/me/_components/settings/settings.tsx
@@ -1,58 +1,16 @@
 'use client'
 
-import React, { useEffect, useState } from 'react'
-import { zodResolver } from '@hookform/resolvers/zod'
-import { Loader2 } from 'lucide-react'
-import { useForm } from 'react-hook-form'
-
-import { updateProfile, UpdateProfilePayload } from '@/validations/profile'
-import { useGetProfile, useUpdateProfile } from '@/hooks/profile/useProfile'
+import React, { useState } from 'react'
+import { useGetProfile } from '@/hooks/profile/useProfile'
 
 import SocialView from '@/sections/me/_components/settings/_components/social'
 
 import PassWordView from './_components/password'
-import Image from 'next/image'
+import ProfileMeView from '@/sections/me/_components/settings/_components/profile'
 
 const MeSetting = () => {
   const [activeTab, setActiveTab] = useState('Profile')
-  const { data: profileData, isLoading } = useGetProfile()
-  const { mutate, isPending } = useUpdateProfile()
-
-  const {
-    register,
-    handleSubmit,
-    setValue,
-    formState: { errors },
-  } = useForm<UpdateProfilePayload>({
-    resolver: zodResolver(updateProfile),
-    defaultValues: {
-      name: '',
-      phone: '',
-      address: '',
-      about_me: '',
-    },
-  })
-
-  useEffect(() => {
-    if (profileData) {
-      setValue('name', profileData?.data.user.name || '')
-      setValue('phone', profileData?.data?.user?.profile?.phone || '')
-      setValue('address', profileData.data?.user?.profile?.address || '')
-      setValue('about_me', profileData.data?.user?.profile?.about_me || '')
-    }
-  }, [profileData, setValue])
-
-  const onSubmit = (data: UpdateProfilePayload) => {
-    mutate(data)
-  }
-
-  if (isLoading) {
-    return (
-      <div className="mt-20">
-        <Loader2 className="mx-auto size-8 animate-spin" />
-      </div>
-    )
-  }
+  const { data: profileData, isLoading: isLoadingProfileData } = useGetProfile()
 
   return (
     <div>
@@ -83,138 +41,19 @@ const MeSetting = () => {
             {/* Tab Content */}
             <div className="widget-content-tab">
               {activeTab === 'Profile' && (
-                <div className="widget-content-inner">
-                  <div className="row">
-                    <div className="profile-wrap">
-                      <div className="profile-img">
-                        <Image
-                          layout="fill"
-                          id="profile-img"
-                          src={profileData?.data.user?.avatar}
-                          alt={profileData?.data.user?.name}
-                        />
-                      </div>
-                      <div className="profile-info">
-                        <h4>{profileData?.data.user?.name}</h4>
-                        <label id="name-file">
-                          PNG or JPG no bigger than 800px wide and tall.
-                        </label>
-                      </div>
-
-                      <div className="profile-btn">
-                        <input id="file-input" type="file" />
-                        <button className="btn-update tf-button-default">
-                          Update <i className="icon-arrow-top-right"></i>
-                        </button>
-                        <a href="#" className="btn-delete tf-button-default">
-                          Delete <i className="icon-arrow-top-right"></i>
-                        </a>
-                      </div>
-                    </div>
-                  </div>
-                  <form
-                    onSubmit={handleSubmit(onSubmit)}
-                    className="shop-checkout"
-                  >
-                    <div className="cols">
-                      <fieldset className="tf-field">
-                        <input
-                          className="tf-input style-1"
-                          id="field1"
-                          type="text"
-                          placeholder="Nhập họ tên của bạn"
-                          tabIndex={2}
-                          {...register('name')}
-                          aria-required="true"
-                        />
-                        <label
-                          className="tf-field-label fs-15"
-                          htmlFor="field1"
-                        >
-                          Họ tên
-                        </label>
-                        {errors.name && (
-                          <span className="text-danger">
-                            {errors.name.message}
-                          </span>
-                        )}
-                      </fieldset>
-                      <fieldset className="tf-field">
-                        <input
-                          className="tf-input style-1"
-                          id="field2"
-                          type="text"
-                          placeholder="Nhập số điện thoại của bạn"
-                          tabIndex={2}
-                          {...register('phone')}
-                        />
-                        <label
-                          className="tf-field-label fs-15"
-                          htmlFor="field2"
-                        >
-                          Số điện thoại
-                        </label>
-                        {errors.phone && (
-                          <span className="text-danger">
-                            {errors.phone.message}
-                          </span>
-                        )}
-                      </fieldset>
-                    </div>
-                    <fieldset className="tf-field mt-4">
-                      <input
-                        className="tf-input style-1"
-                        id="field4"
-                        type="text"
-                        placeholder="Nhập địa chỉ của bạn"
-                        tabIndex={2}
-                        aria-required="true"
-                        {...register('address')}
-                      />
-                      <label className="tf-field-label fs-15" htmlFor="field4">
-                        Địa chỉ
-                      </label>
-                      {errors.address && (
-                        <span className="text-danger">
-                          {errors.address.message}
-                        </span>
-                      )}
-                    </fieldset>
-                    <fieldset className="tf-field mt-4">
-                      <textarea
-                        className="tf-input style-1"
-                        rows={5}
-                        placeholder="Hãy giới thiệu về bạn"
-                        tabIndex={2}
-                        defaultValue="Hãy giới thiệu về bản thân của bạn"
-                        {...register('about_me')}
-                      />
-                      <label
-                        className="tf-field-label type-textarea fs-15"
-                        htmlFor=""
-                      >
-                        Giới thiệu
-                      </label>
-                      {errors.about_me && (
-                        <span className="text-danger">
-                          {errors.about_me.message}
-                        </span>
-                      )}
-                    </fieldset>
-                    <button
-                      type="submit"
-                      className="tf-btn"
-                      disabled={isPending}
-                    >
-                      {isPending && <Loader2 className="animate-spin" />}
-                      Cập nhật hồ sơ <i className="icon-arrow-top-right" />
-                    </button>
-                  </form>
-                </div>
+                <ProfileMeView
+                  profileData={profileData}
+                  isLoadingProfileData={isLoadingProfileData}
+                />
               )}
 
               {activeTab === 'Password' && <PassWordView />}
-              {activeTab === 'Social' && <SocialView />}
+              {activeTab === 'Social' && (
+                <SocialView
+                  profileBioData={profileData}
+                  isLoadingProfileData={isLoadingProfileData}
+                />
+              )}
             </div>
           </div>
         </div>
